fix bugs proceed to mass agg physics

+++ headers/namedefs.h
@@ -2,5 +2,5 @@
 
 namespace wizphys{
 	typedef double real;
-	#define real_pow powf;
+	#define real_pow pow;
 }
